Footer full area done
